+++ src/main/java/nextstep/ladder/controller/LadderGame.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.controller;
 
 import nextstep.ladder.domain.Height;
 import nextstep.ladder.domain.Ladder;
